seed in memory db on startup
