+++ openapi_server/__main__.py
@@ -6,6 +6,7 @@ import connexion
 from openapi_server import encoder
 from flask_cors import CORS
 from openapi_server.database import models
+from openapi_server.database import db_seed
 
 app = connexion.App(__name__, specification_dir="./openapi/")
 app.app.json_encoder = encoder.JSONEncoder
@@ -19,6 +20,7 @@ models.db.init_app(app.app)
 if 'memory' in app.app.config.get("SQLALCHEMY_DATABASE_URI"):
     app.app.app_context().push()
     models.db.create_all()
+    db_seed.seed()
 
 def main():
     app.run(port=8080, debug=True)
